Gets a temperature versus time result from TC-200.
Accomplishes this by using the serial library to send and receive data through COM1. Parses retrieved strings and appends the temperature values with time tags to a list. Uses the Matplotlib library to plot temperature versus time to show stability. Time and number of data points are easily adjustable parameters.